Updating per feedback. Moved RegisterTest tests related to verifyEmail to the new testsuite
